Added Swagger and Added Lambok and map struct
