+++ src/main/java/com/checkpay/controller/UserController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.checkpay.model.User;
+import com.checkpay.model.UserVO;
 import com.checkpay.service.UserService;
 
 @RestController
@@ -15,20 +15,19 @@ public class UserController {
 
 	@Autowired
 	UserService userService;
-	
+
 	@CrossOrigin(origins = "http://localhost:8080")
-    @GetMapping("/api/hello")
-    public String sayHello() {
-        return "Hello from Spring Boot!";
-    }
-	
-	@CrossOrigin(origins = "http://localhost:8080")
-    @PostMapping("/api/user")
-    public String saveUser(@RequestBody User user) {
-		com.checkpay.entity.User userEntity = new com.checkpay.entity.User(user.getPhoneNumber());
-		
-		userService.saveUser(userEntity);
+	@GetMapping("/api/hello")
+	public String sayHello() {
 		return "Hello from Spring Boot!";
-    }
-	
+	}
+
+	@CrossOrigin(origins = "http://localhost:8080")
+	@PostMapping("/api/user")
+	public String saveUser(@RequestBody UserVO user) {
+
+		userService.saveUser(user);
+		return "Hello from Spring Boot!";
+	}
+
 }
