+++ src/main/java/com/checkpay/model/UserVO.java
@@ -1,0 +1,16 @@
+package com.checkpay.model;
+
+import lombok.Data;
+
+@Data
+public class UserVO {
+
+	private String firstName;
+	private String lastName;
+	private String middleName;
+	private Long phoneNumber;
+
+	private String emailId;
+	private Integer pin;
+
+}
