+++ src/main/java/com/checkpay/service/UserService.java
@@ -2,34 +2,16 @@ package com.checkpay.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.checkpay.entity.User;
-import com.checkpay.repository.UserRepository;
+import com.checkpay.model.UserVO;
 
-@Service
-public class UserService {
-	private final UserRepository userRepository;
+public interface UserService {
 
-	@Autowired
-	public UserService(UserRepository userRepository) {
-		this.userRepository = userRepository;
-	}
+	public List<User> getAllUsers();
 
-	public List<User> getAllUsers() {
-		return userRepository.findAll();
-	}
+	public User getUserById(Long id);
 
-	public User getUserById(Long id) {
-		return userRepository.findById(id).orElse(null);
-	}
+	public User saveUser(UserVO user);
 
-	public User saveUser(User user) {
-		return userRepository.save(user);
-	}
-
-	public void deleteUser(Long id) {
-		userRepository.deleteById(id);
-	}
+	public void deleteUser(Long id);
 }
