+++ src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:postgresql://localhost:5432/postgres
+    url: jdbc:postgresql://localhost:5432/Cheqpay
     username: postgres
     password: admin
     driver-class-name: org.postgresql.Driver
