+++ src/main/resources/static/sql/create.sql
@@ -1,19 +1,34 @@
-CREATE TABLE public."user"
+-- Table: public.user
+
+-- DROP TABLE IF EXISTS public."user";
+
+CREATE TABLE IF NOT EXISTS public."user"
 (
-    id bigserial NOT NULL,
-    "firstName" character varying,
-    "lastName" character varying,
-    "middleName" character varying,
-    "phoneNumber" numeric(10) NOT NULL,
-    "emailId" character varying,
-    "pin" numeric(6) NOT NULL,
-    PRIMARY KEY (id)
+    id bigint NOT NULL,
+    email_id character varying(255) COLLATE pg_catalog."default",
+    first_name character varying(255) COLLATE pg_catalog."default",
+    last_name character varying(255) COLLATE pg_catalog."default",
+    middle_name character varying(255) COLLATE pg_catalog."default",
+    phone_number bigint NOT NULL,
+    pin integer NOT NULL,
+    CONSTRAINT user_pkey PRIMARY KEY (id)
 )
 
 TABLESPACE pg_default;
 
 ALTER TABLE IF EXISTS public."user"
     OWNER to postgres;
+    
+-- SEQUENCE: public.user_seq
 
-COMMENT ON TABLE public."user"
-    IS 'Table specific for users';
+-- DROP SEQUENCE IF EXISTS public.user_seq;
+
+CREATE SEQUENCE IF NOT EXISTS public.user_seq
+    INCREMENT 50
+    START 1
+    MINVALUE 1
+    MAXVALUE 9223372036854775807
+    CACHE 1;
+
+ALTER SEQUENCE public.user_seq
+    OWNER TO postgres;
